+++ presentation/src/main/java/com/anytypeio/anytype/presentation/editor/EditorViewModel.kt
@@ -3401,7 +3401,7 @@ class EditorViewModel(
             }
             is ListenerType.Latex -> {
                 when (mode) {
-                    EditorMode.Edit -> proceedWithEnteringActionMode(clicked.id)
+//                    EditorMode.Edit -> proceedWithEnteringActionMode(clicked.id)
                     EditorMode.Select -> onBlockMultiSelectClicked(clicked.id)
                     else -> Unit
                 }